Update chat_interface.py
Corrected chat write to file so the json format is maintained. This allowed me to properly load the chat data from file to the chat window with the current html format.

Note: The selection signal is being sent from the history widget to the main window which in turn is calling on the chat_interface functionality... might not be ideal, but this keeps the logic tightly woven in with the chat_interface widget.