Update chat_interface.py
Wasn't able to get Llamas to work... BUT, I managed to load the current session's data in with the current messages, and now the LLM has conversational context! The chatbot is fully functional! I just had to add in a method to load the currently selected session's data ('readCurrentSessionData') and plug it into the prepareMessage method. Once plugged in, I could hand it back over to my ChatWorker class to add to the payload!