Update history_interface.py
Now a new chat session is automatically created if there are no chat sessions available. Also, the most recent chat session is automatically selected when the user's profile is loaded.

Bug introduced ->
A new chat session will always be automatically generated, even if the user tries to completely delete all sessions. May need to change this in the future.